Reduce renderComponent bundle footprint by ~35% (58KB → 38KB gzipped)
Three changes work together to let bundlers tree-shake heavy deps from
the renderComponent path:

1. Extract renderComponent/BaseRenderer to render-component.ts
   Rollup groups code by source file into shared chunks. Splitting
   renderComponent out of renderer.ts (which also has the classic
   Renderer with routing/outlet/curly deps) produces a separate, small
   shared chunk (~15KB vs ~183KB).

2. Lazy Glimmer global context (glimmer-global-context.ts)
   The setGlobalContext() call previously lived in environment.ts, which
   eagerly imports @ember/-internals/metal, @ember/array, and
   @ember/-internals/views. The new module provides lightweight fallback
   implementations (simple obj[key], Boolean(), etc.) that are replaced
   with full Ember implementations when environment.ts loads during
   classic Ember app boot.

3. Direct import in @ember/renderer
   The public @ember/renderer entry imports directly from
   render-component.ts instead of the @ember/-internals/glimmer barrel,
   so the barrel's other exports don't pull in their deps.

Tested with NullVoxPopuli/smol-est-ember-app (renderComponent only):
  Before: 181KB raw / 58KB gzip
  After:  120KB raw / 38KB gzip

Eliminated from bundle: routing, container, views, service, rsvp.
Reduced: @ember/object+metal 84KB→13KB, @ember/array 18KB→0.2KB.
